Refactor src/components/EventCalendar/EventCalendar.module.css
