+++ src/components/EventCalendar/EventHeader.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Button, Dropdown, Form } from 'react-bootstrap';
 import { Search } from '@mui/icons-material';
-import styles from './EventCalendar.module.css';
+import styles from '../../style/app.module.css';
 import { ViewType } from '../../screens/OrganizationEvents/OrganizationEvents';
 import { useTranslation } from 'react-i18next';
 
@@ -34,7 +34,7 @@ function eventHeader({
   });
 
   return (
-    <div className={styles.calendar}>
+    <div className={styles.calendarEventHeader}>
       <div className={styles.calendar__header}>
         <div className={styles.input}>
           <Form.Control
@@ -55,7 +55,7 @@ function eventHeader({
             onChange={(e) => setEventName(e.target.value)}
           />
           <Button
-            className={styles.searchbutton}
+            className={styles.searchButton}
             style={{ marginBottom: '10px' }}
           >
             <Search />
@@ -64,7 +64,10 @@ function eventHeader({
         <div className={styles.flex_grow}></div>
         <div className={styles.space}>
           <div>
-            <Dropdown onSelect={handleChangeView} className={styles.selectType}>
+            <Dropdown
+              onSelect={handleChangeView}
+              className={styles.selectTypeEventHeader}
+            >
               <Dropdown.Toggle
                 id="dropdown-basic"
                 className={styles.dropdown}
@@ -92,7 +95,7 @@ function eventHeader({
             </Dropdown>
           </div>
           <div>
-            <Dropdown className={styles.selectType}>
+            <Dropdown className={styles.selectTypeEventHeader}>
               <Dropdown.Toggle
                 id="dropdown-basic"
                 className={styles.dropdown}
@@ -112,7 +115,7 @@ function eventHeader({
           </div>
           <Button
             variant="success"
-            className={styles.createButton}
+            className={styles.createButtonEventHeader}
             onClick={showInviteModal}
             data-testid="createEventModalBtn"
           >
